+++ core/vfs_file.cpp
@@ -57,12 +57,12 @@ VfsFile::~VfsFile()
     Close();
 }
 
-Error VfsFile::Read(unsigned long long offset, void* buf, int len)
+Error VfsFile::Read(unsigned long long offset, void* buf, unsigned long len)
 {
     return get_kapi()->vfs_file_read(FilePtr, buf, len, offset);
 }
 
-Error VfsFile::Write(unsigned long long offset, void* buf, int len)
+Error VfsFile::Write(unsigned long long offset, void* buf, unsigned long len)
 {
     return get_kapi()->vfs_file_write(FilePtr, buf, len, offset);
 }
