+++ core/vfs_file.h
@@ -14,8 +14,8 @@ public:
     virtual ~VfsFile();
 
     Error Open(const AString& path, bool read = 0, bool create = 0);
-    Error Read(unsigned long long offset, void* buf, int len);
-    Error Write(unsigned long long offset, void* buf, int len);
+    Error Read(unsigned long long offset, void* buf, unsigned long len);
+    Error Write(unsigned long long offset, void* buf, unsigned long len);
     Error Sync();
     void Close();
